% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_acorn.R
\name{merge_acorn}
\alias{merge_acorn}
\title{Merge the CSV files for every station}
\usage{
merge_acorn(dir)
}
\arguments{
\item{dir}{Directory containing the desired CSV files}
}
\value{
A tibble
}
\description{
Merge the CSV files for every station
}
\examples{
\dontrun{
# merge all stations:
merge_acorn(path/to/directory)
}
}
